requirements.txt
pandas==2.1.4
SQLAlchemy==2.0.24
tabulate==0.9.0
python-dotenv==1.0.0
schedule==1.2.1
pyodbc==5.0.1
mysql-connector-python==8.2.0